Remove private_methods module
